Corrected project
Placed the library files in their own library thingy
