+++ GradleDemo/app/build.gradle
@@ -1,30 +1,17 @@
 /*
  * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java application project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/6.9.3/userguide/building_java_projects.html
  */
 
 plugins {
-    // Apply the application plugin to add support for building a CLI application in Java.
-    id 'application'
-}
-
-repositories {
-    // Use Maven Central for resolving dependencies.
-    mavenCentral()
+    id 'gradledemo.java-application-conventions'
 }
 
 dependencies {
-    // Use JUnit test framework.
-    testImplementation 'junit:junit:4.13'
-
-    // This dependency is used by the application.
-    implementation 'com.google.guava:guava:29.0-jre'
+    implementation 'org.apache.commons:commons-text'
+    implementation project(':utilities')
 }
 
 application {
     // Define the main class for the application.
-    mainClass = 'GradleDemo.App'
+    mainClass = 'gradledemo.app.App'
 }
