+++ GradleDemo/app/src/main/java/gradledemo/app/App.java
@@ -1,0 +1,21 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package gradledemo.app;
+
+import gradledemo.list.LinkedList;
+
+import static gradledemo.utilities.StringUtils.join;
+import static gradledemo.utilities.StringUtils.split;
+import static gradledemo.app.MessageUtils.getMessage;
+
+import org.apache.commons.text.WordUtils;
+
+public class App {
+    public static void main(String[] args) {
+        LinkedList tokens;
+        tokens = split(getMessage());
+        String result = join(tokens);
+        System.out.println(WordUtils.capitalize(result));
+    }
+}
