+++ GradleDemo/app/src/main/java/gradledemo/app/MessageUtils.java
@@ -1,0 +1,10 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package gradledemo.app;
+
+class MessageUtils {
+    public static String getMessage() {
+        return "Hello      World!";
+    }
+}
