+++ GradleDemo/buildSrc/src/main/groovy/gradledemo.java-library-conventions.gradle
@@ -1,0 +1,11 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    // Apply the common convention plugin for shared build configuration between library and application projects.
+    id 'gradledemo.java-common-conventions'
+
+    // Apply the java-library plugin for API and implementation separation.
+    id 'java-library'
+}
