+++ GradleDemo/list/build.gradle
@@ -1,0 +1,7 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    id 'gradledemo.java-library-conventions'
+}
