+++ GradleDemo/settings.gradle
@@ -4,8 +4,8 @@
  * The settings file is used to specify which projects to include in your build.
  *
  * Detailed information about configuring a multi-project build in Gradle can be found
- * in the user manual at https://docs.gradle.org/6.9.3/userguide/multi_project_builds.html
+ * in the user manual at https://docs.gradle.org/7.6/userguide/multi_project_builds.html
  */
 
 rootProject.name = 'GradleDemo'
-include('app')
+include('app', 'list', 'utilities')
