+++ GradleDemo/utilities/build.gradle
@@ -1,0 +1,11 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    id 'gradledemo.java-library-conventions'
+}
+
+dependencies {
+    api project(':list')
+}
