--- Q3Ex11/app/src/main/java/q3ex11/EventWrapper.java
@@ -1,13 +1,0 @@
-package q3ex11;
-
-import javafx.event.Event;
-
-public class EventWrapper {
-    public final Event event;
-    public final String message;
-
-    public EventWrapper(Event event, String message) {
-        this.event = event;
-        this.message = message;
-    }
-}
