--- Q3Ex11/app/src/main/java/q3ex11/ObjectNotFoundException.java
@@ -1,6 +1,0 @@
-package q3ex11;
-
-public class ObjectNotFoundException extends Exception {
-    ObjectNotFoundException() {super();}
-    ObjectNotFoundException(String msg) {super(msg);}
-}
