+++ Q3Ex11/app/src/main/java/q3ex11/app/EventWrapper.java
@@ -1,0 +1,13 @@
+package q3ex11.app;
+
+import javafx.event.Event;
+
+public class EventWrapper {
+    public final Event event;
+    public final String message;
+
+    public EventWrapper(Event event, String message) {
+        this.event = event;
+        this.message = message;
+    }
+}
