+++ Q3Ex11/app/src/main/java/q3ex11/app/ObjectNotFoundException.java
@@ -1,0 +1,6 @@
+package q3ex11.app;
+
+public class ObjectNotFoundException extends Exception {
+    ObjectNotFoundException() {super();}
+    ObjectNotFoundException(String msg) {super(msg);}
+}
