+++ Q3Ex11/buildSrc/src/main/groovy/q3ex11.java-common-conventions.gradle
@@ -1,0 +1,21 @@
+// config file setup for all files?
+
+plugins {
+    id 'java' 
+}
+
+repositories {
+    mavenCentral() 
+}
+
+dependencies {
+    constraints {
+        implementation 'org.apache.commons:commons-text:1.9' 
+    }
+
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.1' 
+}
+
+tasks.named('test') {
+    useJUnitPlatform() 
+}
