+++ Q3Ex11/buildSrc/src/main/groovy/q3ex11.java-library-conventions.gradle
@@ -1,0 +1,5 @@
+plugins {
+    id 'q3ex11.java-common-conventions' 
+
+    id 'java-library' 
+}
